New random filename creation, check if folder exists.
The random generation of filename has been updated to use
rand::distributions::Alphanumeric. This makes it possible to have
repeated characters in filenames, which was what I wanted. The use
of rand::distributions::Alphanumeric means that if someone wants
to customize the characters used, it's going to be slightly harder
(you have to implement the Distribution trait from rand yourself)

There's also now a check that the folder that contains the pastes
exists and tries to create it if it doesn't. If it can't creates
it, the program panics, since it won't be able to function
properly.

Fixes #1